+++ puroro-plugin/src/codegen2/data/field_type.rs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::{Enum, EnumExt, Field, Message, MessageExt, MessageOrEnum, Syntax};
+use super::super::{Enum, Field, Message, MessageOrEnum, Syntax};
 use crate::{ErrorKind, Result};
-use ::proc_macro2::TokenStream;
 use ::puroro_protobuf_compiled::google::protobuf::field_descriptor_proto;
-use ::quote::quote;
 use ::std::rc::Rc;
 
 #[derive(Debug, Clone)]
@@ -126,169 +124,5 @@ impl FieldType {
         } else {
             Err(ErrorKind::MissingTypeName)?
         }
-    }
-
-    pub fn rust_type(&self) -> Result<Rc<TokenStream>> {
-        use FieldType::*;
-        match self {
-            Variant(v) => v.rust_type(),
-            LengthDelimited(ld) => ld.rust_type(),
-            Bits32(b) => b.rust_type(),
-            Bits64(b) => b.rust_type(),
-        }
-    }
-    pub fn rust_maybe_borrowed_type(&self) -> Result<Rc<TokenStream>> {
-        if let FieldType::LengthDelimited(ref ld) = self {
-            ld.rust_maybe_borrowed_type()
-        } else {
-            self.rust_type()
-        }
-    }
-    pub fn rust_mut_ref_type(&self) -> Result<Rc<TokenStream>> {
-        if let FieldType::LengthDelimited(ref ld) = self {
-            ld.rust_mut_ref_type()
-        } else {
-            let raw_type = self.rust_type()?;
-            Ok(Rc::new(quote! { &mut #raw_type }))
-        }
-    }
-    pub fn tag_type(&self) -> Result<Rc<TokenStream>> {
-        use FieldType::*;
-        match self {
-            Variant(v) => v.tag_type(),
-            LengthDelimited(ld) => ld.tag_type(),
-            Bits32(b) => b.tag_type(),
-            Bits64(b) => b.tag_type(),
-        }
-    }
-}
-
-impl VariantType {
-    fn rust_type(&self) -> Result<Rc<TokenStream>> {
-        use VariantType::*;
-        Ok(match self {
-            Int32 | SInt32 => Rc::new(quote! { i32 }),
-            UInt32 => Rc::new(quote! { u32 }),
-            Int64 | SInt64 => Rc::new(quote! { i64 }),
-            UInt64 => Rc::new(quote! { u64 }),
-            Bool => Rc::new(quote! { bool }),
-            Enum2(e) => e.gen_rust_enum_path()?,
-            Enum3(e) => e.gen_rust_enum_path()?,
-        })
-    }
-    fn tag_type(&self) -> Result<Rc<TokenStream>> {
-        use VariantType::*;
-        let tag_name = match self {
-            Int32 => quote! { Int32 },
-            SInt32 => quote! { SInt32 },
-            UInt32 => quote! { UInt32 },
-            Int64 => quote! { Int64 },
-            SInt64 => quote! { SInt64 },
-            UInt64 => quote! { UInt64 },
-            Bool => quote! { Bool },
-            Enum2(e) => {
-                let enum_path = e.gen_rust_enum_path()?;
-                quote! { Enum2 :: <#enum_path> }
-            }
-            Enum3(e) => {
-                let enum_path = e.gen_rust_enum_path()?;
-                quote! { Enum3 :: <#enum_path> }
-            }
-        };
-        Ok(Rc::new(quote! {
-            self::_puroro::tags::#tag_name
-        }))
-    }
-}
-impl LengthDelimitedType {
-    fn rust_type(&self) -> Result<Rc<TokenStream>> {
-        use LengthDelimitedType::*;
-        Ok(match self {
-            String => Rc::new(quote! { ::std::string::String }),
-            Bytes => Rc::new(quote! { ::std::vec::Vec<u8> }),
-            Message(m) => m.gen_rust_struct_path()?,
-        })
-    }
-    fn rust_maybe_borrowed_type(&self) -> Result<Rc<TokenStream>> {
-        use LengthDelimitedType::*;
-        Ok(match self {
-            String => Rc::new(quote! { &str }),
-            Bytes => Rc::new(quote! { &[u8] }),
-            Message(m) => {
-                let path = m.gen_rust_struct_path()?;
-                Rc::new(quote! {
-                    & #path
-                })
-            }
-        })
-    }
-    fn rust_mut_ref_type(&self) -> Result<Rc<TokenStream>> {
-        use LengthDelimitedType::*;
-        Ok(match self {
-            String => Rc::new(quote! { &mut ::std::string::String }),
-            Bytes => Rc::new(quote! { &mut ::std::vec::Vec::<u8> }),
-            Message(m) => {
-                let path = m.gen_rust_struct_path()?;
-                Rc::new(quote! {
-                    &mut #path
-                })
-            }
-        })
-    }
-    fn tag_type(&self) -> Result<Rc<TokenStream>> {
-        use LengthDelimitedType::*;
-        let tag_name = match self {
-            String => quote! { String },
-            Bytes => quote! { Bytes },
-            Message(m) => {
-                let struct_path = m.gen_rust_struct_path()?;
-                quote! { Message :: <#struct_path> }
-            }
-        };
-        Ok(Rc::new(quote! {
-            self::_puroro::tags::#tag_name
-        }))
-    }
-}
-impl Bits32Type {
-    fn rust_type(&self) -> Result<Rc<TokenStream>> {
-        use Bits32Type::*;
-        Ok(Rc::new(match self {
-            Fixed32 => quote! { u32 },
-            SFixed32 => quote! { i32 },
-            Float => quote! { f32 },
-        }))
-    }
-    fn tag_type(&self) -> Result<Rc<TokenStream>> {
-        use Bits32Type::*;
-        let tag_name = match self {
-            Fixed32 => quote! { Fixed32 },
-            SFixed32 => quote! { SFixed32 },
-            Float => quote! { Float },
-        };
-        Ok(Rc::new(quote! {
-            self::_puroro::tags::#tag_name
-        }))
-    }
-}
-impl Bits64Type {
-    fn rust_type(&self) -> Result<Rc<TokenStream>> {
-        use Bits64Type::*;
-        Ok(Rc::new(match self {
-            Fixed64 => quote! { u64 },
-            SFixed64 => quote! { i64 },
-            Double => quote! { f64 },
-        }))
-    }
-    fn tag_type(&self) -> Result<Rc<TokenStream>> {
-        use Bits64Type::*;
-        let tag_name = match self {
-            Fixed64 => quote! { Fixed64 },
-            SFixed64 => quote! { SFixed64 },
-            Double => quote! { Double },
-        };
-        Ok(Rc::new(quote! {
-            self::_puroro::tags::#tag_name
-        }))
     }
 }
