+++ puroro-plugin/src/codegen2/gen.rs
@@ -14,10 +14,12 @@
 
 mod r#enum;
 mod field;
+mod field_type;
 mod message;
 mod package_or_message;
 
 pub use self::r#enum::*;
 pub use self::field::*;
+pub use self::field_type::*;
 pub use self::message::*;
 pub use self::package_or_message::*;
